+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Fixed
 
 - Fixed an issue where the text on the "Add to Itinerary" button was being cut off.
+- Fixed an issue where See all reviews button was not showing the arrow icon.